requirements.txt
# Create and activate your conda env, then run: pip install -r requirements.txt
# Base PyTorch should be installed via conda (cpuonly or CUDA build)
open_clip_torch
transformers
nltk
Pillow
